FEATURE: AccessPoint supports accessPointPaths in the form of http://hostname:port/(path in addition to the existing (port optional) in additional to the standard
port:hostname:path format.

Port can be specified as part of the accessPointPath or, if
its different, can be specified as internalPort property.

This is to allow urls used for replay and query prefix to also
be used for the access point declaration, and allow port to be set differently.

This allows for the access point to be configured using just a single set of
host name properties, even if machine is running on differnt port/hostname