Fix closestGroup race condition
Pull #194 introduced a closestGroup field in RemoteResourceIndex.
However the group object is request-specific and so the values
should not be shared between multiple threads.  The resulting race
condition frequently causes intermittent 404 errors preventing
resources from loading with this message logged:

    Apr 19, 2015 10:52:03 PM org.archive.wayback.webapp.AccessPoint handleReplay
    WARNING: (1)LOADFAIL: Self-Redirect: No Closest Match Found /20140309082106/http://catalogue.nla.gov.au/js/barcode.js?1225315091
    Apr 19, 2015 10:52:03 PM org.archive.wayback.webapp.AccessPoint logError
    WARNING: Runtime Error org.archive.wayback.exception.ResourceNotAvailableException: Self-Redirect: No Closest Match Found

This change replaces the closestGroup field with a local variable
thus ensuring it cannot be overwritten by other threads.