Fix scope and positioning in disclaim.js
Moved into Immediately-Invoked Function Expression to set limit scope and prevent global variables.

Wait until DOM content is loaded to add toolbar (so it doesn't occur prior to body element parsing).

In cases where `background-positions-y`, `padding-top`, or `margin-top` are set on the body element, we'll remove them, remember them in data attributes, and apply them to the bottom of the toolbar to shift the page down. Will restore them to the body when the toolbar is hidden.